+++ Gateway/crds-angular/Services/Interfaces/IPaymentService.cs
@@ -11,8 +11,8 @@ namespace crds_angular.Services.Interfaces
         StripeCustomer GetCustomer(string customerId);
         StripeCustomer DeleteCustomer(string customerId);
         string CreateToken(string accountNumber, string routingNumber);
-        StripeCharge ChargeCustomer(string customerToken, int amount, int donorId);
-        StripeCharge ChargeCustomer(string customerToken, string customerSourceId, int amount, int donorId);
+        StripeCharge ChargeCustomer(string customerToken, decimal amount, int donorId);
+        StripeCharge ChargeCustomer(string customerToken, string customerSourceId, decimal amount, int donorId);
         string UpdateCustomerDescription(string customerToken, int donorId);
         SourceData UpdateCustomerSource(string customerToken, string cardToken);
         SourceData GetDefaultSource(string customerToken);
